+++ src/components/GithubButton.js
@@ -1,9 +1,16 @@
 import styled from 'styled-components';
 import githubLogo from '../assets/images/github-icon-9.png';
 
-export default function GithubButton({ ...props }) {
+export default function GithubButton() {
+  function redirectToGithub() {
+    const GITHUB_URL = 'https://github.com/login/oauth/authorize';
+    const CLIENT_ID = '663f23177c565f1cbd23';
+    const authURL = `${GITHUB_URL}?client_id=${CLIENT_ID}`;
+    window.location.href = authURL;
+  }
+
   return (
-    <GithubButtonContainer {...props}>
+    <GithubButtonContainer onClick={() => redirectToGithub()}>
       <GithubLogo src={githubLogo} alt="github-login" />
       <VerticalDivider></VerticalDivider>
       Login com Github
